# improve performance of parseDVBTime