Using time.Date() instead of time.Parse() and fmt.Sprintf() increases performance by 600%. A benchmark has been added in order to prove it.

Before:

```
cpu: Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
BenchmarkDVBTime
BenchmarkDVBTime-12      3604104               334.9 ns/op
PASS
ok      github.com/asticode/go-astits   1.554s
```

After:

```
cpu: Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
BenchmarkDVBTime
BenchmarkDVBTime-12     22340049                53.35 ns/op
PASS
ok      github.com/asticode/go-astits   1.256s
```